Completed mobile styling and api call
Completed all the mobile styling and added a button animation. Added a delay to set the api call data to give the animation time to complete, and so that the user cant call another piece of advice until at least 2s have passed - can only call the API after a minimum of 2s as the result is cached for that time.